week6/requirements.txt
streamlit
openai
Pillow
python-dotenv
